+++ my-app/app.json
@@ -21,7 +21,6 @@
         "foregroundImage": "./assets/images/adaptive-icon.png",
         "backgroundColor": "#ffffff"
       },
-
       "googleServicesFile": "./google-services.json",
       "package": "com.ucr.agro"
     },
