+++ my-app/app/(tabs)/education.tsx
@@ -1,8 +1,9 @@
+import { theme } from "@/constants/theme";
 import { View, Text } from "react-native";
 
 export default function Education() {
   return (
-    <View>
+    <View style={theme.screenContainer}>
       <Text>Mostrar información para estudiantes</Text>
     </View>
   );
