+++ my-app/app/(tabs)/manageUsers.tsx
@@ -1,8 +1,9 @@
+import { theme } from "@/constants/theme";
 import { View, Text } from "react-native";
 
 export default function ManageUsers() {
   return (
-    <View>
+    <View style={theme.screenContainer}>
       <Text>Mostrar opciones para agregar usuarios</Text>
     </View>
   );
