+++ my-app/app/_layout.tsx
@@ -1,34 +1,38 @@
 import React from "react";
 import { Stack } from "expo-router";
+import { PaperProvider } from "react-native-paper";
+import { theme } from "@/constants/theme";
 
 export default function TabLayout() {
   return (
-    <Stack>
-      <Stack.Screen
-        name="components/FixedVolumeMethod"
-        options={{ headerTitle: "Herbicidas" }}
-      />
-      <Stack.Screen
-        name="components/FixedVelocityMethod"
-        options={{ headerTitle: "Herbicidas" }}
-      />
-      <Stack.Screen
-        name="components/KnownAreaMethod"
-        options={{ headerTitle: "Herbicidas" }}
-      />
-      <Stack.Screen
-        name="components/PesticidePerArea"
-        options={{ headerTitle: "Fungicidas e Insecticidas" }}
-      />
-      <Stack.Screen
-        name="components/PesticidePerPlant"
-        options={{ headerTitle: "Fungicidas e Insecticidas" }}
-      />
-      <Stack.Screen
-        name="components/login/Login"
-        options={{ headerTitle: "Iniciar Sesión" }}
-      />
-      <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-    </Stack>
+    <PaperProvider theme={theme}>
+      <Stack>
+        <Stack.Screen
+          name="components/calculators/FixedVolumeMethod"
+          options={{ headerTitle: "Herbicidas" }}
+        />
+        <Stack.Screen
+          name="components/calculators/FixedVelocityMethod"
+          options={{ headerTitle: "Herbicidas" }}
+        />
+        <Stack.Screen
+          name="components/calculators/KnownAreaMethod"
+          options={{ headerTitle: "Herbicidas" }}
+        />
+        <Stack.Screen
+          name="components/calculators/PesticidePerArea"
+          options={{ headerTitle: "Fungicidas e Insecticidas" }}
+        />
+        <Stack.Screen
+          name="components/calculators/PesticidePerPlant"
+          options={{ headerTitle: "Fungicidas e Insecticidas" }}
+        />
+        <Stack.Screen
+          name="components/login/Login"
+          options={{ headerTitle: "Iniciar Sesión" }}
+        />
+        <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+      </Stack>
+    </PaperProvider>
   );
 }
