+++ my-app/package.json
@@ -13,14 +13,12 @@
     "preset": "jest-expo"
   },
   "dependencies": {
-    "@emotion/react": "^11.11.4",
-    "@expo/vector-icons": "^14.0.1",
-    "@hookform/resolvers": "^3.3.4",
+    "@expo/vector-icons": "^14.0.0",
     "@react-native-firebase/app": "^19.2.2",
     "@react-native-firebase/auth": "^19.2.2",
     "@react-native-firebase/firestore": "^19.2.2",
+    "@hookform/resolvers": "^3.3.4",
     "@react-navigation/native": "^6.0.2",
-    "@theme-ui/components": "^0.16.2",
     "expo": "~50.0.17",
     "expo-font": "~11.10.3",
     "expo-linking": "~6.2.2",
@@ -29,7 +27,7 @@
     "expo-status-bar": "~1.11.1",
     "expo-system-ui": "~2.9.4",
     "expo-web-browser": "~12.8.2",
-    "react": "^18.2.0",
+    "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.51.4",
     "react-native": "0.73.6",
